// src/types/enums/routePath.ts
export enum DefaultRoutePathEnum {
    Home = "/",
    NotFound = "/nao-encontrado",
    RegisterAccount = "/registrar-se",
    Login = "/login",
    RecoverPassword = "/recuperar-senha",
    Profile = "/perfil",
    MyScores = "/pontuacoes/minhas",
    Scores = "/pontuacoes",
    Dashboard = "/dashboard",
    Scenario = "/cenario",
    DecisionsRating = "/cenario/pontuacao"
}